Delete dd.exe on Windows so it is not included in the package.
When UHD is built in release mode, it downloads and includes dd.exe
automatically. However, the download has been failing recently and
breaking link inspection on dd.exe. Also, this should have the license
file included to distribute it properly. So best to avoid these
headaches and not include it in the package.