Set ENABLE_SIM=OFF for all builds.
Having this enabled makes libpython a dependency of libuhd, which almost
doubles the size off the AppImage of Gqrx. Moreover, this apparently
never even worked through conda-forge, because of missing Python
dependencies that the simulator needs to run. Since nobody seems to be
using this and it has a cost to enable, I'll disable it at least for
now.